polc/requirements.txt
Django==1.11
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
PyJWT==1.5.3
Unidecode==1.0.22
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
django-bower==5.2.0
django-braces==1.12.0
django-chartjs==1.2
django-filter==1.1.0
django-highcharts==0.1.7
django-nvd3==0.9.7
djangorestframework==3.7.7
idna==2.6
mongoengine==0.15.0
oauthlib==2.0.6
psycopg2==2.7.4
pymongo==3.6.0
python-decouple==3.1
python-nvd3==0.14.2
python-openid==2.2.5
python-slugify==1.1.4
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
wikipedia==1.4.0
wsgiref==0.1.2
